requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Pillow==10.1.0
google-generativeai==0.3.2
httpx==0.25.2
requests==2.31.0
# Supabase関連
supabase==2.3.4
postgrest==0.13.2
websockets==12.0
